requirements.txt
aiosmtplib==1.1.7
anyio==3.6.2
async-generator==1.10
attrs==22.2.0
bcrypt==4.0.1
black==23.3.0
blinker==1.5
certifi==2023.5.7
click==8.1.3
coverage==7.2.5
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.1.1
fastapi==0.95.1
fastapi-mail==1.1.5
fastapi-pagination==0.11.1
greenlet==2.0.1
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
mypy-extensions==1.0.0
nodeenv==1.8.0
outcome==1.2.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
pluggy==1.0.0
prisma==0.8.2
psycopg2-binary==2.9.5
pydantic==1.10.2
PyJWT==2.6.0
pyright==1.1.309
pyrsistent==0.19.3
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-trio==0.8.0
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.44
sqlalchemy-seeder==0.3.0
starlette==0.26.1
tomli==2.0.1
tomlkit==0.11.8
trio==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
